Fix bug when reading stops writing closes
Instead new behavior is that consumer is allowed to consume until whatever there
is in the buffer to the point producer closed writing.

This is not the case for reading being closed by the consumer, since it doesn't
make sense for producer to keep writing to buffer as it's certain nobody will
consume it, given that this circular buffer implementation is spesifically
tailored for one consumer-one producer case.